[trie] replace legacy uint64 by int

+++ pkg/trie/trie.go
@@ -28,19 +28,19 @@ type Trie struct {
 	// hash is the hash function used in the trie
 	hash func(data ...[]byte) []byte
 	// TrieHeight is the number if bits in a key
-	TrieHeight uint64
+	TrieHeight int
 	// LoadDbCounter counts the nb of db reads in on update
-	LoadDbCounter uint64
+	LoadDbCounter int
 	// loadDbMux is a lock for LoadDbCounter
 	loadDbMux sync.RWMutex
 	// LoadCacheCounter counts the nb of cache reads in on update
-	LoadCacheCounter uint64
+	LoadCacheCounter int
 	// liveCountMux is a lock fo LoadCacheCounter
 	liveCountMux sync.RWMutex
 	// counterOn is used to enable/diseable for efficiency
 	counterOn bool
 	// CacheHeightLimit is the number of tree levels we want to store in cache
-	CacheHeightLimit uint64
+	CacheHeightLimit int
 	// pastTries stores the past maxPastTries trie roots to revert
 	pastTries [][]byte
 	// atomicUpdate, commit all the changes made by intermediate update calls
@@ -51,7 +51,7 @@ type Trie struct {
 func NewTrie(root []byte, hash func(data ...[]byte) []byte, store db.DB) *Trie {
 	s := &Trie{
 		hash:       hash,
-		TrieHeight: uint64(len(hash([]byte("height"))) * 8), // hash any string to get output length
+		TrieHeight: len(hash([]byte("height"))) * 8, // hash any string to get output length
 		counterOn:  false,
 	}
 	s.db = &CacheDB{
@@ -131,7 +131,7 @@ type mresult struct {
 // Adding and deleting can be simultaneous.
 // To delete, set the value to DefaultLeaf.
 // It returns the root of the updated tree.
-func (s *Trie) update(root []byte, keys, values, batch [][]byte, iBatch, height uint64, ch chan<- (mresult)) {
+func (s *Trie) update(root []byte, keys, values, batch [][]byte, iBatch, height int, ch chan<- (mresult)) {
 	if height == 0 {
 		if bytes.Equal(DefaultLeaf, values[0]) {
 			// Delete the key-value from the trie if it is being set to DefaultLeaf
@@ -201,7 +201,7 @@ func (s *Trie) update(root []byte, keys, values, batch [][]byte, iBatch, height 
 }
 
 // updateRight updates the right side of the tree
-func (s *Trie) updateRight(lnode, rnode, root []byte, keys, values, batch [][]byte, iBatch, height uint64, ch chan<- (mresult)) {
+func (s *Trie) updateRight(lnode, rnode, root []byte, keys, values, batch [][]byte, iBatch, height int, ch chan<- (mresult)) {
 	// all the keys go in the right subtree
 	newch := make(chan mresult, 1)
 	s.update(rnode, keys, values, batch, 2*iBatch+2, height-1, newch)
@@ -221,7 +221,7 @@ func (s *Trie) updateRight(lnode, rnode, root []byte, keys, values, batch [][]by
 }
 
 // updateLeft updates the left side of the tree
-func (s *Trie) updateLeft(lnode, rnode, root []byte, keys, values, batch [][]byte, iBatch, height uint64, ch chan<- (mresult)) {
+func (s *Trie) updateLeft(lnode, rnode, root []byte, keys, values, batch [][]byte, iBatch, height int, ch chan<- (mresult)) {
 	// all the keys go in the left subtree
 	newch := make(chan mresult, 1)
 	s.update(lnode, keys, values, batch, 2*iBatch+1, height-1, newch)
@@ -241,7 +241,7 @@ func (s *Trie) updateLeft(lnode, rnode, root []byte, keys, values, batch [][]byt
 }
 
 // updateParallel updates both sides of the trie simultaneously
-func (s *Trie) updateParallel(lnode, rnode, root []byte, lkeys, rkeys, lvalues, rvalues, batch [][]byte, iBatch, height uint64, ch chan<- (mresult)) {
+func (s *Trie) updateParallel(lnode, rnode, root []byte, lkeys, rkeys, lvalues, rvalues, batch [][]byte, iBatch, height int, ch chan<- (mresult)) {
 	lch := make(chan mresult, 1)
 	rch := make(chan mresult, 1)
 	go s.update(lnode, lkeys, lvalues, batch, 2*iBatch+1, height-1, lch)
@@ -268,7 +268,7 @@ func (s *Trie) updateParallel(lnode, rnode, root []byte, lkeys, rkeys, lvalues, 
 }
 
 // deleteOldNode deletes an old node that has been updated
-func (s *Trie) deleteOldNode(root []byte, height uint64, movingUp bool) {
+func (s *Trie) deleteOldNode(root []byte, height int, movingUp bool) {
 	var node Hash
 	copy(node[:], root)
 	if !s.atomicUpdate || movingUp {
@@ -286,7 +286,7 @@ func (s *Trie) deleteOldNode(root []byte, height uint64, movingUp bool) {
 }
 
 // splitKeys devides the array of keys into 2 so they can update left and right branches in parallel
-func (s *Trie) splitKeys(keys [][]byte, height uint64) ([][]byte, [][]byte) {
+func (s *Trie) splitKeys(keys [][]byte, height int) ([][]byte, [][]byte) {
 	for i, key := range keys {
 		if bitIsSet(key, height) {
 			return keys[:i], keys[i:]
@@ -296,7 +296,7 @@ func (s *Trie) splitKeys(keys [][]byte, height uint64) ([][]byte, [][]byte) {
 }
 
 // maybeMoveUpShortcut moves up a shortcut if it's sibling node is default
-func (s *Trie) maybeMoveUpShortcut(left, right, root []byte, batch [][]byte, iBatch, height uint64, ch chan<- (mresult)) bool {
+func (s *Trie) maybeMoveUpShortcut(left, right, root []byte, batch [][]byte, iBatch, height int, ch chan<- (mresult)) bool {
 	if len(left) == 0 && len(right) == 0 {
 		// Both update and sibling are deleted subtrees
 		if iBatch == 0 {
@@ -322,7 +322,7 @@ func (s *Trie) maybeMoveUpShortcut(left, right, root []byte, batch [][]byte, iBa
 	return false
 }
 
-func (s *Trie) moveUpShortcut(shortcut, root []byte, batch [][]byte, iBatch, iShortcut, height uint64, ch chan<- (mresult)) bool {
+func (s *Trie) moveUpShortcut(shortcut, root []byte, batch [][]byte, iBatch, iShortcut, height int, ch chan<- (mresult)) bool {
 	// it doesn't matter if atomic update is true or false since the batch is node modified
 	_, _, shortcutKey, shortcutVal, _, err := s.loadChildren(shortcut, height-1, iShortcut, batch)
 	if err != nil {
@@ -412,7 +412,7 @@ func (s *Trie) maybeAddShortcutToKV(keys, values [][]byte, shortcutKey, shortcut
 
 // loadChildren looks for the children of a node.
 // if the node is not stored in cache, it will be loaded from db.
-func (s *Trie) loadChildren(root []byte, height, iBatch uint64, batch [][]byte) ([][]byte, uint64, []byte, []byte, bool, error) {
+func (s *Trie) loadChildren(root []byte, height, iBatch int, batch [][]byte) ([][]byte, int, []byte, []byte, bool, error) {
 	isShortcut := false
 	if height%4 == 0 {
 		if len(root) == 0 {
@@ -508,7 +508,7 @@ func (s *Trie) parseBatch(val []byte) [][]byte {
 		batch[0] = []byte{0}
 		j := 0
 		for i := 1; i <= 30; i++ {
-			if bitIsSet(bitmap, uint64(i-1)) {
+			if bitIsSet(bitmap, i-1) {
 				batch[i] = val[4+33*j : 4+33*(j+1)]
 				j++
 			}
@@ -519,7 +519,7 @@ func (s *Trie) parseBatch(val []byte) [][]byte {
 
 // leafHash returns the hash of key_value_byte(height) concatenated, stores it in the updatedNodes and maybe in liveCache.
 // leafHash is never called for a default value. Default value should not be stored.
-func (s *Trie) leafHash(key, value, oldRoot []byte, batch [][]byte, iBatch, height uint64) []byte {
+func (s *Trie) leafHash(key, value, oldRoot []byte, batch [][]byte, iBatch, height int) []byte {
 	// byte(height) is here for 2 reasons.
 	// 1- to prevent potential problems with merkle proofs where if an account
 	// has the same address as a node, it would be possible to prove a
@@ -539,7 +539,7 @@ func (s *Trie) leafHash(key, value, oldRoot []byte, batch [][]byte, iBatch, heig
 }
 
 // storeNode stores a batch and deletes the old node from cache
-func (s *Trie) storeNode(batch [][]byte, h, oldRoot []byte, height uint64) {
+func (s *Trie) storeNode(batch [][]byte, h, oldRoot []byte, height int) {
 	if !bytes.Equal(h, oldRoot) {
 		var node Hash
 		copy(node[:], h)
@@ -558,7 +558,7 @@ func (s *Trie) storeNode(batch [][]byte, h, oldRoot []byte, height uint64) {
 }
 
 // interiorHash hashes 2 children to get the parent hash and stores it in the updatedNodes and maybe in liveCache.
-func (s *Trie) interiorHash(left, right, oldRoot []byte, batch [][]byte, iBatch, height uint64) []byte {
+func (s *Trie) interiorHash(left, right, oldRoot []byte, batch [][]byte, iBatch, height int) []byte {
 	var h []byte
 	// left and right cannot both be default. It is handled by moveUpShortcut()
 	if len(left) == 0 {
